requirements.txt
fastapi[all]
uvicorn
python-dotenv
sentence-transformers
PyMuPDF
pinecone
tqdm
requests
